Update display container on quantity bar location
