+++ plugins/tickets/src/modules/blocks/ticket/display-container/display/quantity-bar/__tests__/bar.test.js
