+++ plugins/tickets/src/modules/blocks/ticket/display-container/display/quantity-bar/__tests__/element.test.js
