+++ plugins/tickets/src/modules/blocks/ticket/display-container/display/quantity-bar/bar.js
