+++ plugins/tickets/src/modules/blocks/ticket/display-container/display/quantity-bar/style.pcss
