+++ plugins/tickets/src/modules/blocks/ticket/style.pcss
@@ -22,74 +22,10 @@
 	}
 }
 
-.tribe-editor__ticket-content {
-	flex: 1;
-	padding: 0 26px;
-}
 
-.tribe-editor__ticket-description {
-	color: #545D66;
-	font-size: 0.75rem;
-	font-family: Helvetica, sans-serif, arial;
-	flex: none;
-}
-
-.tribe-editor__ticket-title-container {
-	display: flex;
-	align-items: center;
-
-	button {
-		flex: none;
-	}
-
-	svg {
-		path {
-			fill: #8D949B;
-		}
+.tribe-editor__ticket--display {
+	.tribe-editor__container-panel {
+		width: 100%;
+		flex: 1;
 	}
 }
-
-.tribe-editor__ticket-title {
-	margin: 10px 10px 10px 0;
-	flex: none;
-	line-height: 1;
-}
-
-.tribe-editor__ticket-title,
-.tribe-editor__ticket-price {
-	font-size: 1.3125rem;
-	font-weight: bold;
-	color: #000;
-	font-family: Helvetica, sans-serif, arial;
-}
-
-.tribe-editor__ticket-quantity {
-	padding-right: 25px;
-	flex: none;
-}
-
-.tribe-editor__ticket-price {
-	flex: none;
-	min-width: 65px;
-	margin-right: 20px;
-}
-
-.tribe-editor__quantity-label {
-	color: #000;
-	font-size: 0.9375rem;
-	margin-bottom: 12px;
-	display: inline-block;
-	width: 100%;
-	text-align: center;
-}
-
-.tribe-editor__quantity--unlimited {
-	color: #AEB4BB;
-	font-family: Helvetica, sans-serif, arial;
-	font-size: 0.75rem;
-	line-height: 1;
-	width: 100%;
-	text-align: center;
-	display: inline-block;
-}
-
